+++ migrations/20260619120000_rename_check_ins_comments_to_message.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE check_ins RENAME COLUMN message TO comments;