+++ src/core/ports/mod.rs
@@ -3,5 +3,5 @@ pub mod monitor_repository;
 pub mod notification_dispatcher;
 
 pub use integration_repository::IntegrationRepository;
-pub use monitor_repository::{CheckIn, MonitorRepository};
+pub use monitor_repository::MonitorRepository;
 pub use notification_dispatcher::{NotificationDispatcher, OutboxEntry, OutboxRepository};